[ObjC] Move LineConsumer to its own file.
Make a new (private to the package) target for this so names can depend on it as
an implementation detail as well as the other generator code.

Split the names unittests to have the LineConsumer out on its own also.
